requirements.txt
Django==2.0
django-crispy-forms==1.7.2
pytz==2019.3
